representative done "create statement", in progress "edit statement"

+++ src/main/java/models/entity/Transaction.java
@@ -24,7 +24,7 @@ public class Transaction {
 
     @Getter
     @Setter
-    private int sum;
+    private float sum;
 
 
 }
