+++ src/main/resources/db/migration/V15__Create&Insert_Transaction.sql
@@ -13,7 +13,7 @@ create table trans
       primary key,
   name VARCHAR,
   is_income boolean,
-  sum int
+  sum double precision
 );
 
 insert into trans values
